Break out portions of the monolithic printk patch
- Add
  Revert "tty: serial: meson: Add a earlycon for the T7 SoC"
  serial: Do not hold the port lock when setting rx-during-tx GPIO
  serial: ar933x: Use devm_platform_get_and_ioremap_resource()
  serial: bcm63xx-uart: Use devm_platform_get_and_ioremap_resource()
  serial: clps711x: Use devm_platform_get_and_ioremap_resource()
  serial: core: Fix checks for tx runtime PM state
  serial: core: Fix runtime PM handling for pending tx
  serial: fsl_lpuart: Use devm_platform_get_and_ioremap_resource()
  serial: imx: Use devm_platform_get_and_ioremap_resource()
  serial: linflexuart: Use devm_platform_get_and_ioremap_resource()
  serial: mps2-uart: Use devm_platform_get_and_ioremap_resource()
  serial: mvebu-uart: Use devm_platform_get_and_ioremap_resource()
  serial: omap: Use devm_platform_get_and_ioremap_resource()
  serial: qcom-geni: clean up clock-rate debug printk
  serial: qcom-geni: use icc tag defines
  serial: qcom_geni: Comment use of devm_krealloc rather than devm_krealloc_array
  serial: sccnxp: Use devm_platform_get_and_ioremap_resource()
  serial: sifive: Add suspend and resume operations
  serial: sifive: Remove redundant of_match_ptr()
  serial: sifive: Use devm_platform_get_and_ioremap_resource()
  serial: sprd: Use devm_platform_get_and_ioremap_resource()
  serial: st-asc: Use devm_platform_get_and_ioremap_resource()
  serial: stm32: avoid clearing DMAT bit during transfer
  serial: stm32: do not always set SER_RS485_RX_DURING_TX if RS485 is enabled
  serial: stm32: group dma pause/resume error handling into single function
  serial: stm32: modify parameter and rename stm32_usart_rx_dma_enabled
  serial: stm32: replace access to DMAR bit by dmaengine_pause/resume
  serial: stm32: synchronize RX DMA channel in shutdown
  serial: stm32: use DMAT as a configuration bit
  serial: tegra: Don't print error on probe deferral
  serial: tegra: Use devm_platform_get_and_ioremap_resource()
  serial: vt8500: Use devm_platform_get_and_ioremap_resource()
  tty: serial: fsl_lpuart: add IDLE interrupt support for rx_dma on imx7ulp/imx8ulp/imx8qxp
  tty: serial: fsl_lpuart: move the lpuart32_int() below
  tty: serial: meson: Add a earlycon for the T7 SoC
  tty: serial: meson: add independent uart_data for A1 SoC family
  tty: serial: meson: apply ttyS devname instead of ttyAML for new SoCs
  tty: serial: meson: introduce separate uart_data for S4 SoC family
  tty: serial: meson: redesign the module to platform_driver
  tty: serial: meson: refactor objects definition for different devnames
  tty: serial: meson: use dev_err_probe
  tty: serial: samsung: Set missing PM ops for hibernation support
  tty: serial: samsung_tty: Use abs() to simplify some code
  tty: serial: xilinx_uartps: Do not check for 0 return after calling platform_get_irq()
- Update
  patches.suse/drm-i915-guc-Consider-also-RCU-depth-in-busy-loop.patch
  patches.suse/printk-Update-the-printk-series.patch.
  patches.suse/printk-nbcon-move-locked_port-flag-to-struct-uart_port.patch
  patches.suse/rt-Add-documentation-describing-what-RT-kernel-config-changes-to-default.patch.
  patches.suse/rt-Add-documentation-describing-what-kernel-debug-options-to-add-for-testing.patch.
  patches.suse/rt-Add-helper-script-to-refresh-RT-configs-based-on-the-parent.patch.
  patches.suse/serial-8250-implement-non-BKL-console.patch.
  patches.suse/tty-don-t-check-for-signal_pending-in-send_break.patch.
  patches.suse/tty-use-if-in-send_break-instead-of-goto.patch.
  patches.suse/usb-cdc-acm-return-correct-error-code-on-unsupported-break.patch.

suse-commit: eee7e3baaab4934e7426a57428db67e59b08bb23

+++ include/linux/serial_core.h
@@ -1017,7 +1017,7 @@ void uart_xchar_out(struct uart_port *uport, int offset);
 #ifdef CONFIG_MAGIC_SYSRQ_SERIAL
 #define SYSRQ_TIMEOUT	(HZ * 5)
 
-bool uart_try_toggle_sysrq(struct uart_port *port, u8 ch);
+bool uart_try_toggle_sysrq(struct uart_port *port, unsigned int ch);
 
 static inline int uart_handle_sysrq_char(struct uart_port *port, u8 ch)
 {
